blogpy/requirements.txt
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.3
Django==2.2
django-braces==1.15.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-filter==21.1
django-gravatar2==1.4.4
django-js-asset==1.2.2
django-model-utils==4.2.0
django-star-ratings==0.9.1
django-widget-tweaks==1.4.9
filelock==3.3.2
idna==3.3
Markdown==3.3.4
MarkupSafe==1.1.0
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
swapper==1.2.0
urllib3==1.26.7
virtualenv==20.8.1
